refactor: calendar -> schedule 네이밍 변경
