+++ src/components/schedule/AddSchedule.js
@@ -196,4 +196,3 @@ const AddSchedule = (props) => {
 };
 
 export default AddSchedule;
-
