+++ src/components/schedule/BackgroundPalette.js
