+++ src/pages/reservation/ReservationDetailPage.js
@@ -4,7 +4,6 @@ import { Link } from "react-router-dom";
 
 import ReservationCom from "../../components/reservation/ReservationCom";
 import ReservationDetail from '../../components/reservation/ReservationDetail';
-import Cal from "../../components/calendar/Calendar";
 
 // icon
 import { AiOutlineCalendar } from "react-icons/ai";
